Handling CoAP.SocketServer errors
When the SocketServer attempts to decode messages there is the
possibility of error caused by bad data.

This commit adapts the code so that any exceptions caused by decoding
are captured, and a warning is logged, and the process is stopped
without causing an error.

The types of errors seen were:

- bad return value: {:error, "" is not repeatable""}
- (CondClauseError) no cond clause evaluated to a truthy value (coap_ex
  0.1.0) lib/coap/message_options.ex
- (CaseClauseError) no case clause matching <<...>> (coap_ex 0.1.0)
  lib/coap/message_options.ex:54
- (FunctionClauseError) no function clause matching in
  CoAP.Block.decode/1
- (MatchError) no match of right hand side value: """"

It might be better to update the MessageOptions module to handle these
errors gracefully by throwing a controlled exception. This exception
could then be caught by the SocketServer, allowing it to return a
relevant error along with the partially decoded message. This approach
would enable the SocketServer to potentially identify the connection and
send an appropriate error message to the client.